Added functionality
Added 3 new functions and 2 constants that can be used as variables, but
are protected.  Functions added: 'log', 'log10', 'sqrt'.  Constants
added as used are: '$pi' and '$e'.  Also added new exception define of
'EQEOS_E_NAN' for a non-number being the result of a function - thrown
from log and log10.